Move assertion grant to own file
